indexing_coordinator/requirements.txt
requests
google-cloud-pubsub
protobuf==4.25.6
python-dotenv
